[cscope] update create.sh, add rebuild one module

+++ create_csidx.sh
@@ -3,24 +3,35 @@
 
 DYEL='\E[0;33m'
 DYELL='\E[43;30m'
-DRED='\E[41;30m'
+GRN='\E[1;32m'
 RES='\E[0m'
 
 echo " "
 
 if [ $# -lt 1 ]; then
-    echo "example:"
+    echo -e "${DYELL}example: ${RES}"
     echo " "
-    echo -e "${DYEL}mode 1 single:    ./create_csidx.sh path${RES}"
+
+    echo -e "${DYEL}mode 1: 以目标目录为索引名，创建单个索引  ${RES}"
+    echo -e "${GRN}    build module:    ./create_csidx.sh path ${RES}"
     echo " "
-    echo -e "${DYEL}mode 2 rebuild:   ./create_csidx.sh rebuild${RES}"
+
+    echo -e "${DYEL}mode 2: 以目标目录下的文件夹为索引名，创建多个索引  ${RES}"
+    echo -e "${GRN}    build all:       ./create_csidx.sh path all ${RES}"
     echo " "
-    echo -e "${DYEL}mode 3 all:       ./create_csidx.sh path all${RES}"
+
+    echo -e "${DYEL}mode 3: 更新cscope目录下全部索引  ${RES}"
+    echo -e "${GRN}    rebuild all:     ./create_csidx.sh rebuild ${RES}"
     echo " "
+
+    echo -e "${DYEL}mode 4: 更新cscope目录下单个索引，索引名忽略大小写  ${RES}"
+    echo -e "${GRN}    rebuild module:  ./create_csidx.sh rebuild module ${RES}"
+    echo " "
+
     exit -1
 fi
 
-set -x
+set -e
 FILE_PATH="./file.log"
 REBUILD_FILE="./rebuildfile"
 
@@ -36,7 +47,7 @@ fi
 INPUT_PATH=`cat $FILE_PATH`
 
 if [ "all" == "$2" ]; then
-    echo "mode is all"
+    echo -e "${DYELL} in mode 2, build all ${RES}"
     #rm -rf !(create_csidx.sh)
 
     softfiles=`ls "$INPUT_PATH"`
@@ -51,25 +62,66 @@ if [ "all" == "$2" ]; then
         cscope -bkq -i cscope/"$CSCOPE_FILE".files -f cscope/"$CSCOPE_FILE".out
     done
 elif [ "rebuild" == "$1" ]; then
-    echo "mode is rebuild"
-    rebuildfiles=`ls cscope/*.files`
-    for refiles in ${rebuildfiles}
-    do
-        echo $refiles > $REBUILD_FILE
-        REBUILD_OUT=`sed "s#.files#.out#g" $REBUILD_FILE`
-        TMP_PATH=`sed "s#.files##g" $REBUILD_FILE`
-        echo $TMP_PATH > $REBUILD_FILE
-        TMP_PATH=`sed "s#cscope/##g" $REBUILD_FILE`
-        echo $TMP_PATH > $FILE_PATH
-        REBUILD_PATH=`sed "s#+#/#g" $FILE_PATH`
-        #echo $REBUILD_PATH
-        rm $REBUILD_OUT
-        find "$REBUILD_PATH" -name "*.c" -o -name  "*.cpp" -o -name ".cc" -o -name "*.h" -o -name "*.java" > $refiles
-        cscope -bkq -i $refiles -f $REBUILD_OUT
-    done
-    rm $REBUILD_FILE
+    if [ x"" == x"$2" ]; then
+        echo -e "${DYELL} in mode 3, rebuild all ${RES}"
+        rebuildfiles=`ls cscope/*.files`
+        for refiles in ${rebuildfiles}
+        do
+            echo $refiles > $REBUILD_FILE
+            REBUILD_OUT=`sed "s#.files#.out#g" $REBUILD_FILE`
+            TMP_PATH=`sed "s#.files##g" $REBUILD_FILE`
+            echo $TMP_PATH > $REBUILD_FILE
+            TMP_PATH=`sed "s#cscope/##g" $REBUILD_FILE`
+            echo $TMP_PATH > $FILE_PATH
+            REBUILD_PATH=`sed "s#+#/#g" $FILE_PATH`
+            #echo $REBUILD_PATH
+            rm $REBUILD_OUT
+            find "$REBUILD_PATH" -name "*.c" -o -name  "*.cpp" -o -name ".cc" -o -name "*.h" -o -name "*.java" > $refiles
+            cscope -bkq -i $refiles -f $REBUILD_OUT
+        done
+        rm $REBUILD_FILE
+    else
+        echo -e "${DYELL} in mode 4, rebuild module ${RES}"
+        REBUILD_MODULE=$2
+        echo "rebuild module: $REBUILD_MODULE"
+        rebuildfiles=`ls cscope/*.files`
+        for refiles in ${rebuildfiles}
+        do
+            echo $refiles > $REBUILD_FILE
+            REBUILD_OUT=`sed "s#.files#.out#g" $REBUILD_FILE`
+            TMP_PATH=`sed "s#.files##g" $REBUILD_FILE`
+            echo $TMP_PATH > $REBUILD_FILE
+            TMP_PATH=`sed "s#cscope/##g" $REBUILD_FILE`
+            echo $TMP_PATH > $FILE_PATH
+            REBUILD_PATH=`sed "s#+#/#g" $FILE_PATH`
+            echo $REBUILD_PATH
+            #A=${REBUILD_PATH%/*}/
+            B=${REBUILD_PATH##*/}
+            #echo A:$A
+            #echo B:$B
+
+            shopt -s nocasematch
+            case "$REBUILD_MODULE" in
+                $B ) MATCHED=1;;
+                *) MATCHED=0;;
+            esac
+            shopt -u nocasematch
+
+            if [ x"1" == x"$MATCHED" ]; then
+                echo -e "${DYEL}match, will rebuild ${RES}"
+                rm $REBUILD_OUT
+                find "$REBUILD_PATH" -name "*.c" -o -name  "*.cpp" -o -name ".cc" -o -name "*.h" -o -name "*.java" > $refiles
+                cscope -bkq -i $refiles -f $REBUILD_OUT
+            else
+                echo "not match"
+            fi
+            echo " "
+            
+        done
+
+    fi
 else
-    echo "mode is once"
+    echo -e "${DYELL} in mode 1, build single ${RES}"
     CSCOPE_FILE=`sed "s#/#+#g" $FILE_PATH`
 
     find "$INPUT_PATH" -name "*.c" -o -name  "*.cpp" -o -name ".cc" -o -name "*.h" -o -name "*.java" > cscope/"$CSCOPE_FILE".files
